Добавлен основной метод в фабрике, а также метод для разделения строки с параметрами объекта на отдельные параметры
