+++ Laba1/factory.h
@@ -1,10 +1,11 @@
 #ifndef UTILSH
 
 #include <string> 
+#include "Menu.h"
 
 using namespace std;
 
-vector<shared_ptr<Menu>> foo(vector<string> inputStrings){}
+vector<shared_ptr<Menu>> foo(vector<string> inputStrings);
 
 #endif
